Fixed TSV converter and Leave_main_sample

+++ source/variants/filtering.py
@@ -550,8 +550,11 @@ def postprocess_vcf(sample, original_anno_vcf_fpath, work_filt_vcf_fpath):
     # Converting to TSV
     if work_filt_vcf_fpath and 'tsv_fields' in cnf:
         tsv_fpath = make_tsv(cnf, work_filt_vcf_fpath, sample.name)
-
-        if isfile(final_tsv_fpath): os.remove(final_tsv_fpath)
+        if not tsv_fpath:
+            err('TSV convertion didn\'t work for ' + sample.name)
+            final_tsv_fpath = None
+        else:
+            if isfile(final_tsv_fpath): os.remove(final_tsv_fpath)
         shutil.copy(tsv_fpath, final_tsv_fpath)
     else:
         final_tsv_fpath = None
